+++ CHANGELOG.md
@@ -29,6 +29,9 @@ release.
   tensor context and the `jStartProof`, `jStopProof`, `jIntros`, `jDestruct`,
   `jPoseProof`, `jAssert`, `jSplitL`, `jSplitR`, `jApply`, and
   `jUnitIntro` tactics
+- Remove legacy tactics for `ctx_refines` composition. (`ctxr_refl`,
+  `ctxr_transL`, `ctxr_transR`, `ctxr_norm`, `ctxr_swap`, `ctxr_rotate`,
+  `ctxr_drop`)
 
 ## 2026-07-22